backend/requirements.txt
fastapi==0.104.1
motor==3.3.2
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
python-multipart==0.0.6
